+++ common/src/main/java/io/github/mortuusars/envelope/event/CommonEvents.java
@@ -15,9 +15,6 @@ public class CommonEvents {
     }
 
     public static void levelTick(Level level) {
-        if (level instanceof ServerLevel serverLevel) {
-            serverLevel.getEnvelopeContext().tick();
-        }
     }
 
     /**