+++ common/src/main/java/io/github/mortuusars/envelope/world/service/KnownPlayers.java
@@ -11,6 +11,7 @@ import net.minecraft.nbt.NbtOps;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.level.saveddata.SavedData;
+import net.minecraft.world.level.storage.DimensionDataStorage;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
@@ -89,8 +90,8 @@ public class KnownPlayers extends SavedData {
 
     // -- Save / Load
 
-    public static KnownPlayers get(ServerLevel level) {
-        return level.getServer().overworld().getDataStorage().computeIfAbsent(KnownPlayers.factory(), "envelope_known_players");
+    public static KnownPlayers get(ServerLevel level, String name) {
+        return level.getDataStorage().computeIfAbsent(factory(), name);
     }
 
     @Override